+++ src/main/java/org/example/Main.java
@@ -1,7 +1,6 @@
 package org.example;
 
-import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.*;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
@@ -17,22 +16,5 @@ public class Main {
             System.out.println(s);
         }
          */
-        for (int i=0; i<10; i++) {
-            System.out.println(fibonacci(i));
-        }
-
-    }
-
-    public static int fibonacci(int n) {
-        if (n <= 1) return n;
-        int prev = 0;
-        int current = 1;
-        for (int i=2; i<=n; i++)
-        {
-            int temp = prev + current;
-            prev = current;
-            current = temp;
-        }
-        return current;
     }
 }
